Update permalinks.py
I have removed the .encode('utf-8') in the article_url() function because it appends b' to the string of the URL and as a result, the URL redirection does not work.
Ps. I am using Python 3.5.2 in Windows 10 and I am a beginner... (just joined to GitHub 10min ago to propose this change...). Why you use .encode('utf-8') ?
Ps. I have found interesting also this explanation... http://stackoverflow.com/questions/6269765/what-does-the-b-character-do-in-front-of-a-string-literal